+++ modules/versions.lua
@@ -6,6 +6,8 @@ return {
   ["Gen1BattleUI"] = "1.5.2",
   ["Gen1BillsBox"] = "1.5.0",
   ["Gen1Dex"] = "1.5.0",
+  ["Gen1Elevator"] = "maintained",
+  ["Gen1ItemInfo"] = "maintained",
   ["Gen1MenuManager"] = "0.2.8",
   ["Gen1ModMenu"] = "0.9.0",
   ["Gen1ModernBag"] = "1.9.4",